[refactoring] hand를 초기화 하는 역할 추가
