+++ src/main/java/blackjack/domain/rule/HandInitializer.java
@@ -1,0 +1,24 @@
+package blackjack.domain.rule;
+
+import blackjack.domain.deck.Deck;
+import blackjack.domain.gamer.Dealer;
+import blackjack.domain.gamer.Player;
+import blackjack.domain.gamer.Players;
+
+public class HandInitializer {
+
+    private static final int INITIAL_HAND_COUNT = 2;
+
+    public static void initialize(Dealer dealer, Players players, Deck deck) {
+        for (int i = 0; i < INITIAL_HAND_COUNT; i++) {
+            drawCard(dealer, players, deck);
+        }
+    }
+
+    private static void drawCard(Dealer dealer, Players players, Deck deck) {
+        dealer.draw(deck.pick());
+        for (Player player : players) {
+            player.draw(deck.pick());
+        }
+    }
+}
